refactor: melhorias gerais nos componentes
