+++ src/components/LoginForm/index.tsx
@@ -3,22 +3,49 @@ import { getApiKey } from '../../redux/actions/apiKey';
 
 import './style.scss';
 import { Button } from '../__common__/Button';
-import { Error } from '../__common__/Error';
+import { Error as ErrorComponent } from '../__common__/Error';
 import { Link } from '../__common__/Link';
 import { useApi } from '../../hooks/useApi';
-import { handleLogin } from './functions/handleLogin';
 import { useNavigate } from 'react-router';
 import { Form } from '../__common__/Form';
 import { useAuth } from '../../core/context/AuthContext';
+import { erro, success } from '../../utils/toatsFunctions';
+import { useEffect, useRef } from 'react';
 
 interface ApiKeyProps {
   apiKey: any;
 }
 
 const LoginForm = ({ apiKey }: ApiKeyProps): JSX.Element => {
+  const inputRef: any = useRef();
+
+  useEffect(() => {
+    if (inputRef.current) {
+      inputRef.current.focus();
+    }
+  }, []);
+
   const api = useApi();
   const navigate = useNavigate();
   const { login } = useAuth();
+
+  const handleLogin = async (api: any) => {
+    try {
+      const data = await api.login();
+      success(
+        `Login efetuado com sucesso ${data.response.account.firstname}! `,
+      );
+      return data;
+    } catch (error) {
+      const errorMessage: HTMLElement | HTMLSpanElement | any =
+        document.querySelector('.u-iserror');
+      errorMessage.style.display = 'block';
+      erro(
+        `Desculpe algo saiu errado, verifique sua chave de acesso e tente novamente.`,
+      );
+      throw new Error('Erro no login');
+    }
+  };
 
   const handleSubmit = async () => {
     const userData = await handleLogin(api);
@@ -36,21 +63,18 @@ const LoginForm = ({ apiKey }: ApiKeyProps): JSX.Element => {
           <p className="c-login_title">Meu Time</p>
         </div>
         <div className="c-logincard__form">
-          {/* <Input
-            type="password"
-            placeholder="Insira sua chave de acesso aqui..."
-            onChange={apiKey}
-          >
-            Chave de acesso
-          </Input> */}
           <label className="c-logincard__label">Chave de Acesso</label>
           <input
+            ref={inputRef}
             className="c-logincard__input"
             type="password"
             placeholder="ex: 1234567890"
             onChange={e => apiKey(e.target.value)}
           />
-          <Error> Chave inválida ou inexistente. Tente novamente. </Error>
+          <ErrorComponent>
+            {' '}
+            Chave inválida ou inexistente. Tente novamente.{' '}
+          </ErrorComponent>
         </div>
         <Button type="submit">ENTRAR</Button>
       </Form>
