--- src/components/__common__/Input/style.scss
@@ -1,20 +1,0 @@
-.c-logincard__input {
-  background-color: transparent;
-  border: 1px solid #bebebe;
-  border-radius: 8px;
-  height: fit-content;
-  width: 100%;
-  padding: 12px 8px;
-  font-size: 1rem;
-
-  &::placeholder {
-    color: #bebebe;
-    font-size: 0.8rem;
-  }
-}
-
-.c-logincard__label {
-  font-size: 0.7rem;
-  padding-bottom: 4px;
-  font-weight: normal;
-}
